tt-outreach-agent/requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
anthropic==0.25.0
python-dotenv==1.0.0
